Day_028_Pomodoro/requirements.txt
Pillow==9.3.0
pyttsx3==2.90